Interface ISnapshottable instead of abstract class
